% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTreePrior.R
\name{E_alpha}
\alias{E_alpha}
\title{Expected value of the number of bottom nodes in the unrealistic case where we assume that the number of variables and possible splits are infinite (therefore P(T) is not dependent on the design matrix X) and \eqn{\beta=0} (Case #1).}
\usage{
E_alpha(alpha)
}
\arguments{
\item{alpha}{base parameter of the tree prior, \eqn{alpha \in [0,1)}.}
}
\value{
Returns the expected value of the number of bottom nodes.
}
\description{
Expected value of the number of bottom nodes in the unrealistic case where we assume that the number of variables and possible splits are infinite (therefore P(T) is not dependent on the design matrix X) and \eqn{\beta=0} (Case #1).
}
\examples{
E_alpha(.30)
E_alpha(.45)
E_alpha(.499)
E_alpha(.75)
}
\references{
Jolicoeur-Martineau, A. (Currently in revision, expected 2016) \emph{Etude d'une loi a priori pour les arbres binaires de regression} (\emph{Study on the prior distribution of binary regression trees}) (Master thesis). UQAM university.
}
\seealso{
\code{\link{Var_alpha}}
}

