% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTreePrior.R
\name{GetListUniqueSplits}
\alias{GetListUniqueSplits}
\title{Unique splits that leads to children with more than \eqn{minpart} nodes.}
\usage{
GetListUniqueSplits(x, minpart = 1)
}
\arguments{
\item{x}{vector containing the observations of a variable}

\item{minpart}{minimum number of observations in the children nodes}
}
\value{
Returns a vector containing the unique splits that leads to children with more than \eqn{minpart} nodes.
}
\description{
Unique splits that leads to children with more than \eqn{minpart} nodes.
}
\examples{
GetListUniqueSplits(c(1,4,7,3,0,2,2,3,4,7,7,7),minpart=1)
GetListUniqueSplits(c(1,4,7,3,0,2,2,3,4,7,7,7),minpart=3)
}

