% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTreePrior.R
\name{NumBotMaxDepth_inf}
\alias{NumBotMaxDepth_inf}
\title{Number of bottom nodes and depth in the unrealistic case where we assume that the number of variables and possible splits are infinite (therefore P(T) is not dependent on the design matrix X) (Case #2).}
\usage{
NumBotMaxDepth_inf(alpha, beta, depth = 0)
}
\arguments{
\item{alpha}{base parameter of the tree prior, \eqn{\alpha \in [0,1)}.}

\item{beta}{power parameter of the tree prior, \eqn{beta \geq 0}.}

\item{depth}{depth of the current node, \eqn{depth \geq 0}.}
}
\value{
Returns a vector containing the number of bottom nodes and depth.
}
\description{
Generate a tree and returns the number of bottom nodes and depth in the unrealistic case where we assume that the number of variables and possible splits are infinite (therefore P(T) is not dependent on the design matrix X) (Case #2).
}
\examples{
NumBotMaxDepth_inf(.95,.5)
}
\seealso{
\code{\link{NumBotMaxDepth}}, \code{\link{NumBotMaxDepthX}}
}

