% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesTreePrior.R
\name{p_split}
\alias{p_split}
\title{Probability of split of the tree prior.}
\usage{
p_split(alpha, beta, depth = 0)
}
\arguments{
\item{alpha}{base parameter of the tree prior, \eqn{\alpha \in [0,1)}.}

\item{beta}{power parameter of the tree prior, \eqn{\beta \geq 0}.}

\item{depth}{depth of the current node, \eqn{depth \geq 0}.}
}
\value{
Returns the probability of split.
}
\description{
Probability of split of the tree prior.
}
\examples{
p_split(.95,.5)
p_split(.95,.5,1)
p_split(.95,.5,2)
}
\references{
Chipman, H. A., George, E. I. et McCulloch, R. E. (1998). Bayesian CART model search. Journal of the American Statistical Association, 93(443), 935-948.
}

